Change generateQuestion function to class QuestionGenerator
